(culling) Extract FilterBar and ContextMenu subcomponents
Move the filter/action bar and the right-click context menu JSX verbatim
into components/culling/FilterBar.jsx and ContextMenu.jsx as presentational
components. The shell renders them with narrow props; behavior unchanged.
